+++ zvt/domain/stats/__init__.py
@@ -1,1 +1,3 @@
 # -*- coding: utf-8 -*-
+from .stock_1d_ma_stats import *
+from .stock_1wk_ma_stats import *
