+++ zvt/recorders/eastmoney/finance/china_stock_finance_factor_recorder.py
@@ -154,6 +154,6 @@ class ChinaStockFinanceFactorRecorder(BaseChinaStockFinanceRecorder):
 
 
 if __name__ == '__main__':
-    # init_process_log('finance_factor.log')
+    # init_log('finance_factor.log')
     recorder = ChinaStockFinanceFactorRecorder(codes=['000001'])
     recorder.run()
