+++ zvt/recorders/eastmoney/trading_runner.py
@@ -7,7 +7,7 @@ from apscheduler.schedulers.background import BackgroundScheduler
 
 from zvt.recorders.eastmoney.trading.holder_trading_recorder import HolderTradingRecorder
 from zvt.recorders.eastmoney.trading.manager_trading_recorder import ManagerTradingRecorder
-from zvdata.utils.utils import init_process_log
+from zvt import init_log
 
 logger = logging.getLogger(__name__)
 
@@ -29,7 +29,7 @@ def run():
 
 
 if __name__ == '__main__':
-    init_process_log('trading.log')
+    init_log('trading.log')
 
     run()
