+++ zvt/recorders/joinquant/joinquant_stock_bar_runner.py
@@ -5,7 +5,7 @@ import time
 from apscheduler.schedulers.background import BackgroundScheduler
 
 from zvdata import IntervalLevel
-from zvdata.utils.utils import init_process_log
+from zvt import init_log
 from zvt.recorders.joinquant.quotes.jq_stock_bar_recorder import JQChinaStockBarRecorder
 
 logger = logging.getLogger(__name__)
@@ -28,7 +28,7 @@ def run():
 
 
 if __name__ == '__main__':
-    init_process_log('joinquant_bar_runner.log')
+    init_log('joinquant_bar_runner.log')
 
     run()
