+++ zvt/recorders/sina/sina_runner.py
@@ -4,9 +4,9 @@ import time
 
 from apscheduler.schedulers.background import BackgroundScheduler
 
+from zvt import init_log
 from zvt.recorders.sina.meta.sina_china_stock_category_recorder import SinaChinaStockCategoryRecorder
 from zvt.recorders.sina.money_flow.sina_index_money_flow_recorder import SinaIndexMoneyFlowRecorder
-from zvdata.utils.utils import init_process_log
 
 logger = logging.getLogger(__name__)
 
@@ -27,7 +27,7 @@ def run():
 
 
 if __name__ == '__main__':
-    init_process_log('sina_runner.log')
+    init_log('sina_runner.log')
 
     run()
